% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GLM-methods.R
\docType{methods}
\name{dimension}
\alias{dimension}
\alias{dimension,GLM-method}
\title{dimension method}
\usage{
dimension(object)

\S4method{dimension}{GLM}(object)
}
\arguments{
\item{object}{one of \code{GLM} object}
}
\value{
dimension

list
        N predictor number for each observation
        P feature number in GLM
        K class number of observation
}
\description{
dimension method

dimension method return the dimension of variable X
             {n x p | n = #observation, p = #feature, k = #class}
}

