% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Logreg-methods.R
\docType{methods}
\name{error,logisticRegression-method}
\alias{error,logisticRegression-method}
\title{error calculation}
\usage{
\S4method{error}{logisticRegression}(object, model)
}
\description{
error calculation
}

