% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{error}
\alias{error}
\title{error method}
\usage{
error(object, model)
}
\arguments{
\item{object}{one of \code{GLM} object}
}
\value{
error object
}
\description{
error method
}

