% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataClasses.R
\docType{class}
\name{GLM-class}
\alias{GLM-class}
\title{__cplusplus__ class nested && S4 object nested}
\description{
GLM-class is a virtual Class. Inherited by C++ backend object or S4 object
}
\details{
Class "GLM"
}
\section{Slots}{

\describe{
\item{\code{lambda}}{L2 regulizer parameter}
}}
\author{
xin zhou \url{xxz220@miami.edu}
}
\keyword{classes}

