% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-fit.R
\name{glm_fit}
\alias{glm_fit}
\title{Fit a GLM}
\usage{
glm_fit(X, y, family, maxit = 1000, tol = 1e-10)
}
\arguments{
\item{X}{a design matrix}

\item{y}{the response}

\item{family}{the family}

\item{maxit}{maximum number of iteration}

\item{tol}{tolerance for convergence}
}
\description{
Fit a GLM
}
