% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Benchmark.R
\name{glmCV}
\alias{glmCV}
\title{glmCV}
\usage{
glmCV(X, Y, lambda = 10^(seq(-3, 1, by = 0.5)), fold = 5)
}
\arguments{
\item{lambda}{numeric vector}

\item{fold}{cross validation fold}

\item{data}{}
}
\value{
best lambda
}
\description{
function of logistic regression CV
}
\examples{
data(mini_mnist)
glmCV(X=mini_mnist$train$images, Y=mini_mnist$train$labels)
}
\author{
xin zhou \url{xxz220@miami.edu}
}

