% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Logreg-methods.R
\docType{methods}
\name{gradient}
\alias{gradient}
\alias{gradient,logisticRegression-method}
\title{gradient method}
\usage{
gradient(object, w)

\S4method{gradient}{logisticRegression}(object, w)
}
\arguments{
\item{object}{one of \code{logisticRegression} object}

\item{w}{coefficient in Xw}
}
\value{
summary object

gradient matrix
}
\description{
gradient method

return gradient of logistic regression model
}

