% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Optimizer-methods.R
\docType{methods}
\name{initialize,Optimizer-method}
\alias{initialize,Optimizer-method}
\title{constructor for Optimizer}
\usage{
\S4method{initialize}{Optimizer}(.Object, p, w = NULL,
  direction = c("gradient", "newton", "stochastic gradient"),
  stepsize = c("wolfe", "backtrack"), wtol = 1e-07, gtol = 1e-07,
  ftol = 1e-07, iter = 100, t = 1, ...)
}
\arguments{
\item{p}{optimization problem class}

\item{direction}{character c("gradient", "newton", "stochastic gradient")}

\item{stepsize}{character c("wolfe", "backtrack")}

\item{wtol}{coefficient criteria for stop iteration}

\item{gtol}{gradient tolerance}

\item{ftol}{function value tolerance}

\item{w0}{initialization coefficient matrix}

\item{maxiter}{maximum iteration; default = 100}
}
\description{
constructor for Optimizer
}
\examples{
data(mini_mnist)
LR <- new("logisticRegression", X=mini_mnist$train$images, Y=mini_mnist$train$labels, lambda=0.5)
O  <- new("Optimizer", p = LR, w = matrix(0, nrow = 784, ncol = 10), iter = 1000)
}

