% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Loadmnist.R
\name{labelMNIST}
\alias{labelMNIST}
\title{labelMNIST}
\usage{
labelMNIST(Y)
}
\arguments{
\item{Y}{output matrix}
}
\value{
label of digits, from 0 to 9
}
\description{
label_mnist : transform Y(n x K) to digits label
}
\author{
xin zhou \url{xxz220@miami.edu}
}

