% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Loadmnist.R
\name{loadMNIST}
\alias{loadMNIST}
\title{loadMNIST}
\usage{
loadMNIST(verbose = TRUE)
}
\arguments{
\item{verbose}{bool; Default TRUE}
}
\value{
List
        train
        test
}
\description{
Load MNIST digit recognition database(.Rda) into R environment
}
\author{
xin zhou \url{xxz220@miami.edu}
}

