% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataClasses.R
\docType{class}
\name{logisticRegression-class}
\alias{logisticRegression-class}
\title{optimization related function
Class logisticRegression}
\description{
optimization related function
Class logisticRegression
}
\section{Slots}{

\describe{
\item{\code{X}}{x values (bias term added) [1, x]}

\item{\code{Y}}{y values Matrix}

\item{\code{w}}{coefficient values}

\item{\code{lambda}}{parameter for regularization}
}}
\author{
Zhou Xin
}
\seealso{
\code{\link{logisticRegression}}
}
\keyword{classes}

