% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Logreg-methods.R
\name{loglikelihood}
\alias{loglikelihood}
\title{loglikelihood to calculate the objval for logistic regression}
\usage{
loglikelihood(y, X, w)
}
\description{
loglikelihood to calculate the objval for logistic regression
}

