% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppGLM-package.R
\docType{data}
\name{mini_mnist}
\alias{mini_mnist}
\title{mini_mnist data is the hand writing digits classification tasks}
\format{mini_mnist : list}
\usage{
data(mini_mnist, package = 'RcppGLM')
}
\description{
mini_mnist data is the hand writing digits classification tasks
}
\keyword{datasets}

