% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Benchmark.R
\name{mm_benchmark}
\alias{mm_benchmark}
\title{mm_benchmark}
\usage{
mm_benchmark(size = 10)
}
\arguments{
\item{size}{scale of tested matrix; default 10}
}
\description{
function of benchmark subset
}
\author{
xin zhou \url{xxz220@miami.edu}
}

