% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Loadmnist.R
\name{mnist_show}
\alias{mnist_show}
\title{mnist_show}
\usage{
mnist_show(img, grid = 28, ncol = 10, plot = TRUE, label)
}
\arguments{
\item{img}{mnist image matrix}

\item{grid}{picture pixel resolution}

\item{ncol}{digit number each row}
}
\description{
plot logistic regression training set
}
\details{
mnist_show
}
\examples{
imag <- loadMNIST()
train <- imag$train$images[(1:100), ]
mnist_show(train)
}
\author{
xin zhou \url{xxz220@miami.edu}
}

