% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataClasses.R
\docType{class}
\name{Optimizer-class}
\alias{Optimizer-class}
\title{Class "Optimizer"}
\description{
Class "Optimizer"
}
\section{Slots}{

\describe{
\item{\code{p}}{\code{GLM} object}

\item{\code{w}}{coefficient matrix}

\item{\code{w_prev}}{coefficient matrix of last iteration}

\item{\code{g}}{gradient matrix}

\item{\code{g_prev}}{gradient matrix of last iteration}

\item{\code{f}}{objective function value}

\item{\code{f_prev}}{objective function value of last iteration}

\item{\code{d}}{direction coefficient matrix}

\item{\code{direction}}{direction methods; character c("gradient", "newton", "stochastic gradient")}

\item{\code{stepsize}}{stepsize methods; character c("wolfe", "backtrack")}

\item{\code{maxiter}}{maximum iteration; default = 100}

\item{\code{iter}}{iteration number}

\item{\code{wtol}}{coefficient criteria for stop iteration}

\item{\code{gtol}}{gradient tolerance}

\item{\code{ftol}}{function value tolerance}

\item{\code{t}}{stepsize}
}}
\author{
xin zhou \url{xxz220@miami.edu}
}
\keyword{classes}

