% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Logreg-methods.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,logisticRegression-method}
\title{predict method}
\usage{
predict(object, model)

\S4method{predict}{logisticRegression}(object, model)
}
\arguments{
\item{object}{one of \code{GLM}, \code{logisticRegression} object}

\item{model}{optimizer model}

\item{object}{input observation matrix}
}
\value{
predict object
}
\description{
predict method

predict categorical of each observation
}

