% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppGLM-package.R
\docType{package}
\name{RcppGLM}
\alias{RcppGLM}
\alias{RcppGLM-package}
\alias{package-RcppGLM}
\title{A parallel generalization linear model solver for large scale data}
\description{
This package implements versions of generalization linear model that are
appropriate for learning big data. It does so by wrapping the (C++) library
developed by Eric Xin Zhou.
}

