% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Logreg-methods.R
\docType{methods}
\name{solve_hessian}
\alias{solve_hessian}
\alias{solve_hessian,logisticRegression-method}
\title{solve_hessian method}
\usage{
solve_hessian(object, w, g)

\S4method{solve_hessian}{logisticRegression}(object, w, g)
}
\arguments{
\item{object}{one of \code{logisticRegression} object}

\item{w}{coefficient in Xw}

\item{g}{gradient of matrix}
}
\value{
solve_hessian object
}
\description{
solve_hessian method

return solved direction of newton rasphon (H^{-1}g)
}

