% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{summary}
\alias{summary}
\title{summary method}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{one of \code{Optimizer} object}
}
\value{
summary object
}
\description{
summary method
}

