% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Optimizer-methods.R
\docType{methods}
\name{train}
\alias{train}
\alias{train,Optimizer-method}
\title{train method}
\usage{
train(object, verbose = TRUE)

\S4method{train}{Optimizer}(object, verbose = TRUE)
}
\arguments{
\item{object}{one of \code{Optimizer} object}

\item{verbose}{display iteration information; default TRUE}
}
\value{
train object
}
\description{
train method

Optimizer train
}

